separate role module
